requirements.txt
Flask>=2.0
flask-cors
paho-mqtt
numpy
scipy
matplotlib
